+++ Anime.h
@@ -12,14 +12,14 @@ class Anime
 public:
 	int id;
 	string name;
-	int rating;
+	float rating;
 	vector<string> genres;
 	int episodes;
     float normEpisodes; // normalized value for episodes
 
 	Anime();
 	~Anime();
-	Anime(int id, string name, int rating, vector<string> genres, int episodes, float normEpisodes);
+	Anime(int id, string name, float rating, vector<string> genres, int episodes, float normEpisodes);
 
 
 };
